Separate pipelines and use Ninja for builds

+++ test/control/LoadHandlerTest.cpp
@@ -78,10 +78,10 @@ public:
 		LoadHandler handler;
 		Document* doc = handler.loadDocument(GET_TESTFILE("test1.xoj"));
 
-		CPPUNIT_ASSERT_EQUAL(1UL, doc->getPageCount());
+		CPPUNIT_ASSERT_EQUAL((size_t) 1, doc->getPageCount());
 		PageRef page = doc->getPage(0);
 
-		CPPUNIT_ASSERT_EQUAL(1UL, (*page).getLayerCount());
+		CPPUNIT_ASSERT_EQUAL((size_t) 1, (*page).getLayerCount());
 		Layer* layer = (*(*page).getLayers())[0];
 
 		Element* element = (*layer->getElements())[0];
@@ -97,10 +97,10 @@ public:
 		LoadHandler handler;
 		Document* doc = handler.loadDocument(GET_TESTFILE("packaged_xopp/test.xopp"));
 
-		CPPUNIT_ASSERT_EQUAL(1UL, doc->getPageCount());
+		CPPUNIT_ASSERT_EQUAL((size_t) 1, doc->getPageCount());
 		PageRef page = doc->getPage(0);
 
-		CPPUNIT_ASSERT_EQUAL(1UL, (*page).getLayerCount());
+		CPPUNIT_ASSERT_EQUAL((size_t) 1, (*page).getLayerCount());
 		Layer* layer = (*(*page).getLayers())[0];
 
 		Element* element = (*layer->getElements())[0];
@@ -116,10 +116,10 @@ public:
 		LoadHandler handler;
 		Document* doc = handler.loadDocument(GET_TESTFILE("test1.unzipped.xoj"));
 
-		CPPUNIT_ASSERT_EQUAL(1UL, doc->getPageCount());
+		CPPUNIT_ASSERT_EQUAL((size_t) 1, doc->getPageCount());
 		PageRef page = doc->getPage(0);
 
-		CPPUNIT_ASSERT_EQUAL(1UL, (*page).getLayerCount());
+		CPPUNIT_ASSERT_EQUAL((size_t) 1, (*page).getLayerCount());
 		Layer* layer = (*(*page).getLayers())[0];
 
 		Element* element = (*layer->getElements())[0];
@@ -135,7 +135,7 @@ public:
 		LoadHandler handler;
 		Document* doc = handler.loadDocument(GET_TESTFILE("load/pages.xoj"));
 
-		CPPUNIT_ASSERT_EQUAL(5UL, doc->getPageCount());
+		CPPUNIT_ASSERT_EQUAL((size_t) 5, doc->getPageCount());
 	}
 
 	void testPagesZipped()
@@ -143,7 +143,7 @@ public:
 		LoadHandler handler;
 		Document* doc = handler.loadDocument(GET_TESTFILE("packaged_xopp/pages.xopp"));
 
-		CPPUNIT_ASSERT_EQUAL(5UL, doc->getPageCount());
+		CPPUNIT_ASSERT_EQUAL((size_t) 5, doc->getPageCount());
 	}
 
 	void checkPageType(Document* doc, int pageIndex, string expectedText, PageType expectedBgType)
@@ -153,7 +153,7 @@ public:
 		PageType bgType = page->getBackgroundType();
 		CPPUNIT_ASSERT(expectedBgType == bgType);
 
-		CPPUNIT_ASSERT_EQUAL(1UL, (*page).getLayerCount());
+		CPPUNIT_ASSERT_EQUAL((size_t) 1, (*page).getLayerCount());
 		Layer* layer = (*(*page).getLayers())[0];
 
 		Element* element = (*layer->getElements())[0];
@@ -168,7 +168,7 @@ public:
 		LoadHandler handler;
 		Document* doc = handler.loadDocument(GET_TESTFILE("load/pages.xoj"));
 
-		CPPUNIT_ASSERT_EQUAL(5UL, doc->getPageCount());
+		CPPUNIT_ASSERT_EQUAL((size_t) 5, doc->getPageCount());
 		checkPageType(doc, 0, "p1", PageType("plain"));
 		checkPageType(doc, 1, "p2", PageType("lined"));
 		checkPageType(doc, 2, "p3", PageType("ruled"));
@@ -181,7 +181,7 @@ public:
 		LoadHandler handler;
 		Document* doc = handler.loadDocument(GET_TESTFILE("packaged_xopp/pages.xopp"));
 
-		CPPUNIT_ASSERT_EQUAL(5UL, doc->getPageCount());
+		CPPUNIT_ASSERT_EQUAL((size_t) 5, doc->getPageCount());
 		checkPageType(doc, 0, "p1", PageType("plain"));
 		checkPageType(doc, 1, "p2", PageType("lined"));
 		checkPageType(doc, 2, "p3", PageType("ruled"));
@@ -205,10 +205,10 @@ public:
 		LoadHandler handler;
 		Document* doc = handler.loadDocument(GET_TESTFILE("load/layer.xoj"));
 
-		CPPUNIT_ASSERT_EQUAL(1UL, doc->getPageCount());
+		CPPUNIT_ASSERT_EQUAL((size_t) 1, doc->getPageCount());
 		PageRef page = doc->getPage(0);
 
-		CPPUNIT_ASSERT_EQUAL(3UL, (*page).getLayerCount());
+		CPPUNIT_ASSERT_EQUAL((size_t) 3, (*page).getLayerCount());
 		checkLayer(page, 0, "l1");
 		checkLayer(page, 1, "l2");
 		checkLayer(page, 2, "l3");
@@ -219,10 +219,10 @@ public:
 		LoadHandler handler;
 		Document* doc = handler.loadDocument(GET_TESTFILE("packaged_xopp/layer.xopp"));
 
-		CPPUNIT_ASSERT_EQUAL(1UL, doc->getPageCount());
+		CPPUNIT_ASSERT_EQUAL((size_t) 1, doc->getPageCount());
 		PageRef page = doc->getPage(0);
 
-		CPPUNIT_ASSERT_EQUAL(3UL, (*page).getLayerCount());
+		CPPUNIT_ASSERT_EQUAL((size_t) 3, (*page).getLayerCount());
 		checkLayer(page, 0, "l1");
 		checkLayer(page, 1, "l2");
 		checkLayer(page, 2, "l3");
@@ -233,10 +233,10 @@ public:
 		LoadHandler handler;
 		Document* doc = handler.loadDocument(GET_TESTFILE("load/text.xml"));
 
-		CPPUNIT_ASSERT_EQUAL(1UL, doc->getPageCount());
+		CPPUNIT_ASSERT_EQUAL((size_t) 1, doc->getPageCount());
 		PageRef page = doc->getPage(0);
 
-		CPPUNIT_ASSERT_EQUAL(1UL, (*page).getLayerCount());
+		CPPUNIT_ASSERT_EQUAL((size_t) 1, (*page).getLayerCount());
 		Layer* layer = (*(*page).getLayers())[0];
 
 		Text* t1 = (Text*)(*layer->getElements())[0];
@@ -262,10 +262,10 @@ public:
 		LoadHandler handler;
 		Document* doc = handler.loadDocument(GET_TESTFILE("packaged_xopp/text.xopp"));
 
-		CPPUNIT_ASSERT_EQUAL(1UL, doc->getPageCount());
+		CPPUNIT_ASSERT_EQUAL((size_t) 1, doc->getPageCount());
 		PageRef page = doc->getPage(0);
 
-		CPPUNIT_ASSERT_EQUAL(1UL, (*page).getLayerCount());
+		CPPUNIT_ASSERT_EQUAL((size_t) 1, (*page).getLayerCount());
 		Layer* layer = (*(*page).getLayers())[0];
 
 		Text* t1 = (Text*)(*layer->getElements())[0];
